net: l2: ieee802154: Change the hardware filters while associating
During an association, the peer we are trying to connect to will send us
an association response frame with the destination PAN ID set to the PAN
ID we try to join. If we do not update the hardware address filters
beforehands, it is likely that the hardware will just discard the
response and the association will fail.

(cherry picked from commit 31b35973e3811e5bef46374fc37e7c279ec1c2aa)

Original-Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>
GitOrigin-RevId: 31b35973e3811e5bef46374fc37e7c279ec1c2aa
Change-Id: I377152daaa71704f0b7994fd4c6239fdda048e0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4131000
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>