net: l2: ieee802154: Fix a wrong channel change during association
The current association logic starts by changing the channel. The way it
is done is wrong because it dereferences req->channel which is simply
not initialized by the caller. But anyway, the command itself does not
support providing a channel so we must already be on the right one when
trying to associate. Hence, drop this channel change call.

(cherry picked from commit 6a11e79b4b3edf2d40940ead6acee0326105dc14)

Original-Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>
GitOrigin-RevId: 6a11e79b4b3edf2d40940ead6acee0326105dc14
Change-Id: I763490623edc2441f89f87fc1ea2f8847f971c0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4131001
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>