Bluetooth: Mesh: Add CDB depend on to prov Kconfig
The current implementation of provisioning features is dependent on CDB
to compile, which is reflected poorly in the Kconfig options for mesh.

This commit alters the Kconfig options for provisioning to be dependent
on the CDB option to compile.

(cherry picked from commit 36e34fce41db5efa0ef4175358dc539ae691a416)

Original-Signed-off-by: Anders Storrø <anders.storro@nordicsemi.no>
GitOrigin-RevId: 36e34fce41db5efa0ef4175358dc539ae691a416
Change-Id: Ib48fa7a8d2334ce666f6d1ea99035f97d85b2cab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3882679
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>