Bluetooth: Shell: Add connect-name command
Add a new command, connect-name, which takes a string
as the single input and uses that to setup a scan filter
and connect to the first device that matches the name.

This command is thus basically shorthand for
bt scan-filter-set <name>
bt scan on
bt scan off
bt connect
bt scan-filter-clear-name

The idea of this command is to not only make it easier
find and connect to devices in a crowded environment,
but also to more easily connect to devices that changes
their advertised address.

The implementation is minimal as it uses existing
shell commands.

(cherry picked from commit 34d4501e8a7b0686e99dfa21df1aa7f4cfe2103e)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 34d4501e8a7b0686e99dfa21df1aa7f4cfe2103e
Change-Id: If42757f9cea05ea0c51928502505f4eac9f413a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3886767
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>