drivers: clock_control: stm32 fixed LSE system clock enable
On some stm32 mcus, the LSE is enabled as system clock (LSESYS)
only when the LSEON and LSERDY are both set.
The bit LSESYSEN is set in the RCC BDCR register
and the driver is waiting for the LSESYSRDY to be set.

(cherry picked from commit 1816e0c7ea3bce463f67bebebe98ecd2cb99ba35)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: 1816e0c7ea3bce463f67bebebe98ecd2cb99ba35
Change-Id: I9690cd8b120ad6189d25664b11bd37f8ecb9167c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3886771
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>