kernel/sched: fix condition for CPU mask set
When building with CONFIG_SCHED_CPU_MASK_PIN_ONLY=y, CPU mask
is fixed and cannot be changed while thread is running.

The current code asserts if thread state is anything but PREPARED.

We do however have interface like k_work_queue_start() where a thread is
started as part of the queue start. To allow user to set the pinned CPU
for the work queue thread, it needs to be possible to suspend the
thread, set the mask, and then call k_thread_resume(). This seems to be
a valid sequence, so relax the assert check to reflect this.

(cherry picked from commit e81ccef613a6c0c50deac90effe729abc672c449)

Original-Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
GitOrigin-RevId: e81ccef613a6c0c50deac90effe729abc672c449
Change-Id: I786627983547e9a92a80c3401799c1d96ce548a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3889546
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>