soc: xtensa: tools: do not interfere with IPC in logs-only mode
The tool has caused some hard to detect failures when used
concurrently with the SOF Linux driver to observe logs. This
has been rootcaused to the code to handle and send IPC messages
that has been active even in logs-only mode. In vast majority
of cases, the SOF kernel driver is faster to act and handles
the IPC interrupts. But occasionally the cavstool.py timing was
just right and it managed to ack a message before Linux driver.

Fix the problem by not handling IPC messages when the tool
is run in logs-only mode.

(cherry picked from commit 1a6d8684007d754b6b876701117e693613831465)

Original-Reported-by: Peter Ujfalusi <peter.ujfalusi@linux.intel.com>
Original-Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
GitOrigin-RevId: 1a6d8684007d754b6b876701117e693613831465
Change-Id: I1bf001a49a8bf0cddb26760363a65fb6dd601448
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3891260
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>