includes: prevent warnings (Z_CBPRINTF_ARG_SIZE shadows variables)
We compile with the gcc compiler flags -Werror and -Wshadow. So
the mentioned macro rasies a warning (and with the flags the
warning becoms to an error). It uses the same variable names
as on of its caller. Add an addidional underscore to the
affected variables to prevent this warning.

(cherry picked from commit 387d84e4525e93fefee893029b50cd055c6d6cb1)

Original-Signed-off-by: Peter Tönz <peter.tonz@husqvarnagroup.com>
GitOrigin-RevId: 387d84e4525e93fefee893029b50cd055c6d6cb1
Change-Id: I40c3f8d13940f0bc0f0f9cfab19b6c5a60aee219
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3921667
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>