Bluetooth: l2cap: reword unknown pdu warning
This is a scary-looking warning for users, and should only really matter to
stack developers. It will eventually show up in the wild for embedded
devices that don't get updated to the latest and greatest.

Reword it to make it clear we are rejecting it, and it doesn't impact the
stack in any way.

(cherry picked from commit 477c0fcb6e9e0899144e035985d43933de11234d)

Original-Signed-off-by: Jonathan Rico <jonathan.rico@nordicsemi.no>
GitOrigin-RevId: 477c0fcb6e9e0899144e035985d43933de11234d
Change-Id: I2b5e6ab21d7268cc8fca449bf68d4a57840a5c16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3934382
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>