Bluetooth: Audio: Clear received_base in audio shell on PA sync lost
When the PA sync is lost, we clear the received_base so
that once we resync to the PA, we get the BASE again.

This is to make it easier to decode and use the BASE again
if the sync was lost, as the previous BASE may have
been printed a long time before.

(cherry picked from commit 208f44722f699ab456e9299017f07518c5864fbc)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 208f44722f699ab456e9299017f07518c5864fbc
Change-Id: Ia748ae5312f51056c274d93dc976f696dc477606
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3939358
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>