device: s/dev_name/dev_id
The usage of dev_name is misleading, since it does not represent the
device name (struct device name field) but the name given to the defined
struct device. In practice, it is used as a kind of unique device
identtifier, so let's rename it to dev_id.

(cherry picked from commit 33161222d7632bf6d9074e72007217ebc1669667)

Original-Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: 33161222d7632bf6d9074e72007217ebc1669667
Change-Id: I9e8d4c0163d5b5fbc2b4e79fd1a2f55d33405705
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3934593
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>