Bluetooth: Audio: Fix segmentation issue in BT Audio shell
It was possible to enter a state where setting the QoS
for a source stream after having released a sink stream
as the unicast client would cause a segmentation error.
This issue is related to how the unicast shell is using
the unicast group.

A fix has been added for it, and a check for
whether or not `audio init` (to register the needed
callback) has been added. This check has also been added
for the broadcast sink.

(cherry picked from commit e4b640ff2b06c6278f96260051d237cbc687ffe1)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: e4b640ff2b06c6278f96260051d237cbc687ffe1
Change-Id: I536cbe81ad2c643ac8229cb1f67f7f119e7688d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3958516
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>