Bluetooth: Audio: Ensure that stream is in streaming state for ISO recv
Adds check to verify that the audio stream is actually
in the streaming state before handling the
audio data. As per the BAP spec, a stream that
is not in the streaming state shall not send
any data, and this check is to handle any
remote devices that are not working correctly.

(cherry picked from commit 4321fef3b70a836eee4dfa1c53ead78662f8f3b2)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 4321fef3b70a836eee4dfa1c53ead78662f8f3b2
Change-Id: Ib5c5cb302a29ce82cb5560aac2478a790f404467
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3966791
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>