drivers: timer: use sys_read64 to read HPET counter on 64 bits cpu
For 32 bit processor to read the 64 bits hpet counter, the HPET spec
2.4.7 suggest to read HPET counter high and low then checking the
high bits to decide if it rollovers or not.

But this logic seems to cause problem for 64 bits processor under SMP,
there is a possible one tick earier under tickless mode. It is likely
to be the cache coherence issue, because a mfence instruction before
reading the timer works.

So we change to read the 64 bits counter by sys_read64 on 64bit
processor to prevent this issue.

Fixes #49611

(cherry picked from commit dcda15d17d86d6c8fd3947d980a0607ae86c9719)

Original-Signed-off-by: Enjia Mai <enjia.mai@intel.com>
GitOrigin-RevId: dcda15d17d86d6c8fd3947d980a0607ae86c9719
Change-Id: I138bd6e6f3dfc22dba177549864cbf993c5bc221
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3948995
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>