drivers: ipm: fix infinite TXF interrupt
The TXF interrupt mask is not properly set. In this case the
interrupt line is not cleaned, leading to the re-entrance
in the interrupt handler.

the issue is linked to the management of the IPCC_CxTOCySR value.
The TXF interrupt is triggered when the IPCC_CyMR is channel bit is
set to 0 and the IPCC_CxTOCySR associated bit move from 1 to 0.

The fix is to check if the IPCC_CxTOCySR channel bit is 0 instead of 1.

(cherry picked from commit fb43a480651fcaeb9360e00ea905b871775e1005)

Original-Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
GitOrigin-RevId: fb43a480651fcaeb9360e00ea905b871775e1005
Change-Id: I4396d8726b8361df9e941bd5d0780c432913e823
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3966790
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>