tests/kernel/workq: Fix semaphore counting
This test case will call k_sem_give() twice and expect both to be
received by k_sem_take(), yet the semaphore is initialized with a
maximum count of one!

The reason this worked was an undocumented misfeature of k_sem: if
k_sem_take() was called on a semaphore with a pended thread, it would
wake up that thread synchronously instead of incrementing the count.
So you could call it once to wake up the thread and again to queue the
count and not overflow.  The problem is that this is a priority bug (a
high priority runnable thread should have the chance to run and call
k_sem_take() before a low priority thread that got woken).

Zync corrects that, and so needs to have two slots if you want two
semaphore events.

(cherry picked from commit 6a1f721dda1cadf243459c6b554de90f9458af4c)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 6a1f721dda1cadf243459c6b554de90f9458af4c
Change-Id: I725efefc2c3673a89cf730fd8fa03b33a1e6d039
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3960034
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>